+++ src/fields/EscapeDamField.php
@@ -12,8 +12,10 @@ use craft\errors\VolumeException;
 use craft\fields\Assets;
 use craft\helpers\Assets as AssetsHelper;
 use craft\helpers\FileHelper;
+use craft\models\VolumeFolder;
 
-use \RuntimeException;
+use escape\escapedam\EscapeDam;
+
 use yii\base\InvalidConfigException;
 
 class EscapeDamField extends Assets
@@ -47,9 +49,6 @@ class EscapeDamField extends Assets
      */
     public bool $enableAssetsInput = true;
 
-    // Static
-    // =========================================================================
-
     /**
      * @inheritdoc
      */
@@ -58,9 +57,12 @@ class EscapeDamField extends Assets
         if ((\debug_backtrace()[1]['function'] ?? null) === 'twig_get_attribute') {
             return Craft::t('app', 'Assets');
         }
-        return Craft::t('app', 'Escape DAM');
+        return EscapeDam::getInstance()->getSettings()->pluginName ?? Craft::t('app', 'Escape DAM');
     }
 
+    /**
+     * @return string
+     */
     public static function defaultDamSelectionLabel(): string
     {
         return Craft::t('escapedam', 'Import from DAM');
@@ -86,15 +88,25 @@ class EscapeDamField extends Assets
     }
 
     /**
-     * Resolve source path for DAM importing for this field.
-     *
-     * @param ElementInterface|null $element
-     * @throws InvalidSubpathException
-     * @throws VolumeException
+     * @inheritdoc
      */
-    public function resolveDynamicPathToImportFolderId(ElementInterface $element = null): int
+    public function getSettingsHtml(): ?string
     {
-        return $this->_determineImportFolderId($element, true);
+        $this->damImportLocationSource = $this->_volumeSourceToFolderSource($this->damImportLocationSource);
+        return parent::getSettingsHtml();
+    }
+
+    /**
+     * @param ElementInterface|null $element
+     * @return int
+     * @throws InvalidSubpathException
+     * @throws InvalidVolumeException
+     * @throws VolumeException
+     * @throws \ReflectionException
+     */
+    public function getImportFolderId(ElementInterface $element = null): int
+    {
+        return $this->_importFolder($element)->id;
     }
 
     /**
@@ -117,6 +129,51 @@ class EscapeDamField extends Assets
             'damSelectionLabel' => $this->damSelectionLabel ?: self::defaultDamSelectionLabel(),
             'allowedExtensions' => $allowedExtensions,
         ]);
+    }
+
+    /**
+     * Determine an upload folder id by looking at the settings and whether Element this field belongs to is new or not.
+     *
+     * @param ElementInterface|null $element
+     * @return VolumeFolder
+     * @throws InvalidSubpathException
+     * @throws InvalidVolumeException
+     * @throws VolumeException
+     */
+    private function _importFolder(ElementInterface $element = null): VolumeFolder
+    {
+        /** @var Element $element */
+        $sourceKey = $this->damImportLocationSource;
+        $subpath = $this->damImportLocationSubpath;
+        if (!$subpath || !strlen($subpath)) {
+            $subpath = strtolower($this->handle) . '-' . $this->id . '/' . date('Ymdhis');
+        }
+        $settingName = Craft::t('escapedam', 'Import Location');
+        $assets = Craft::$app->getAssets();
+        try {
+            if (!$sourceKey || !$folder = $this->_findFolder($sourceKey, $subpath, $element, true)) {
+                throw new VolumeException();
+            }
+        } catch (VolumeException $e) {
+            throw new InvalidVolumeException(Craft::t('app', 'The {field} field’s {setting} setting is set to an invalid volume.', [
+                'field' => $this->name,
+                'setting' => $settingName,
+            ]), 0, $e);
+        } catch (InvalidSubpathException $e) {
+            // If this is a new/disabled element, the subpath probably just contained a token that returned null, like {id}
+            // so use the user's upload folder instead
+            if ($element === null || !$element->id || !$element->enabled) {
+                $folder = $assets->getUserTemporaryUploadFolder();
+            } else {
+                // Existing element, so this is just a bad subpath
+                throw new InvalidSubpathException($e->subpath, Craft::t('app', 'The {field} field’s {setting} setting has an invalid subpath (“{subpath}”).', [
+                    'field' => $this->name,
+                    'setting' => $settingName,
+                    'subpath' => $e->subpath ?? '',
+                ]), 0, $e);
+            }
+        }
+        return $folder;
     }
 
     /**
@@ -150,111 +207,36 @@ class EscapeDamField extends Assets
     }
 
     /**
-     * Determine an upload folder id by looking at the settings and whether Element this field belongs to is new or not.
-     *
-     * @param ElementInterface|null $element
-     * @param bool $createDynamicFolders
-     * @return int
-     * @throws VolumeException
+     * @param string|null $sourceKey
+     * @return string
      */
-    private function _determineImportFolderId(ElementInterface $element = null, bool $createDynamicFolders = true): int
+    private function _volumeSourceToFolderSource(?string $sourceKey = null): string
     {
-        /** @var Element $element */
-        $importVolume = $this->damImportLocationSource;
-        $subpath = $this->damImportLocationSubpath;
-        if (!$subpath) {
-            $subpath = \strtolower($this->handle) . '-' . $this->id . '/' . date('Ymdhis');
-        }
-        $settingName = Craft::t('escapedam', 'Import Location');
-        $assets = Craft::$app->getAssets();
-        try {
-            if (!$importVolume) {
-                throw new VolumeException();
-            }
-            $folderId = $this->_resolveVolumePathToFolderId($importVolume, $subpath, $element, $createDynamicFolders);
-        } catch (VolumeException $e) {
-            throw new VolumeException(Craft::t('app', 'The {field} field’s {setting} setting is set to an invalid volume.', [
-                'field' => $this->name,
-                'setting' => $settingName,
-            ]), 0, $e);
-        } catch (\Throwable $e) {
-            // If this is a new/disabled element, the subpath probably just contained a token that returned null, like {id}
-            // so use the user's upload folder instead
-            if ($element === null || !$element->getId() || !$element->enabled || !$createDynamicFolders) {
-                $userFolder = $assets->getUserTemporaryUploadFolder();
-                $folderId = $userFolder->id;
-            } else {
-                // Existing element, so this is just a bad subpath
-                throw new RuntimeException(Craft::t('app', 'The {field} field’s {setting} setting has an invalid subpath (“{subpath}”).', [
-                    'field' => $this->name,
-                    'setting' => $settingName,
-                    'subpath' => $subpath,
-                ]), 0, $e);
+        if ($sourceKey && is_string($sourceKey) && strpos($sourceKey, 'volume:') === 0) {
+            $parts = explode(':', $sourceKey);
+            $volume = Craft::$app->getVolumes()->getVolumeByUid($parts[1]);
+
+            if ($volume && $folder = Craft::$app->getAssets()->getRootFolderByVolumeId($volume->id)) {
+                return 'folder:' . $folder->uid;
             }
         }
-        return $folderId;
+
+        return (string)$sourceKey;
     }
 
     /**
-     * Resolve a source path to its folder ID by the source path and the matched source beginning.
-     *
-     * @param string $uploadSource
-     * @param string|null $subpath
+     * @param string $sourceKey
+     * @param string $subpath
      * @param ElementInterface|null $element
      * @param bool $createDynamicFolders
-     * @return int
-     * @throws VolumeException
-     * @throws \Throwable
-     * @throws \yii\base\Exception
+     * @return VolumeFolder
+     * @throws \ReflectionException
      */
-    private function _resolveVolumePathToFolderId(string $uploadSource, ?string $subpath = '', ElementInterface $element = null, bool $createDynamicFolders = true): int
+    private function _findFolder(string $sourceKey, string $subpath, ElementInterface $element = null, bool $createDynamicFolders = true): VolumeFolder
     {
-        $assetsService = Craft::$app->getAssets();
-        $volumeId = $this->_volumeIdBySourceKey($uploadSource);
-        // Make sure the volume and root folder actually exists
-        if ($volumeId === null || ($rootFolder = $assetsService->getRootFolderByVolumeId($volumeId)) === null) {
-            throw new VolumeException();
-        }
-        // Are we looking for a subfolder?
-        $subpath = is_string($subpath) ? trim($subpath, '/') : '';
-        if ($subpath === '') {
-            // Get the root folder in the source
-            $folderId = $rootFolder->id;
-        } else {
-            // Prepare the path by parsing tokens and normalizing slashes.
-            $renderedSubpath = Craft::$app->getView()->renderObjectTemplate($subpath, $element);
-            // Did any of the tokens return null?
-            if (
-                $renderedSubpath === '' ||
-                trim($renderedSubpath, '/') != $renderedSubpath ||
-                str_contains($renderedSubpath, '//')
-            ) {
-                throw new RuntimeException("Subpath \"$subpath\" contains null tokens");
-            }
-            // Sanitize the subpath
-            $segments = explode('/', (string)$renderedSubpath);
-            foreach ($segments as &$segment) {
-                $segment = FileHelper::sanitizeFilename($segment, [
-                    'asciiOnly' => Craft::$app->getConfig()->getGeneral()->convertFilenamesToAscii
-                ]);
-            }
-            unset($segment);
-            $subpath = implode('/', $segments);
-            $folder = $assetsService->findFolder([
-                'volumeId' => $volumeId,
-                'path' => $subpath . '/'
-            ]);
-            // Ensure that the folder exists
-            if (!$folder) {
-                if (!$createDynamicFolders) {
-                    throw new RuntimeException("Unable to find or create folder for subpath \"$subpath\"");
-                }
-                $volume = Craft::$app->getVolumes()->getVolumeById($volumeId);
-                $folderId = $assetsService->ensureFolderByFullPathAndVolume($subpath, $volume)->id;
-            } else {
-                $folderId = $folder->id;
-            }
-        }
-        return $folderId;
+        $method = new \ReflectionMethod(parent::class, '_findFolder');
+        $method->setAccessible(true);
+        return $method->invokeArgs($this, [$sourceKey, $subpath, $element, $createDynamicFolders]);
     }
+
 }