+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Added 
+- Adds Craft 4.4 compatibility
+### Changed 
+- Escape DAM now requires Craft 4.4 or later
+- It's no longer possible to select volumes in the `allowedVolumesForImport` settings as "Native Assets" sources   
+
 ## 2.1.0
 ### Added
 - Adds a "VIDEO" tag to DAM video thumbnails